+++ test/integration/test_build.py
@@ -6,7 +6,10 @@ Many of the tests will depend on a "builddir" fixture which is a compiled wily c
 
 TODO : Test build + build with extra operator
 """
+import os
+import sys
 import pathlib
+import tempfile
 import pytest
 from click.testing import CliRunner
 from git import Repo, Actor
@@ -14,45 +17,63 @@ from mock import patch
 
 import wily.__main__ as main
 from wily.archivers import ALL_ARCHIVERS
+from wily.config import generate_cache_path
+
+_path = "src\\test.py" if sys.platform == "win32" else "src/test.py"
 
 
-def test_build_not_git_repo(tmpdir):
+def test_build_not_git_repo(tmpdir, cache_path):
     """
     Test that build defaults to filesystem in a non-git directory
     """
-    with patch("wily.logger") as logger:
-        runner = CliRunner()
-        result = runner.invoke(main.cli, ["--path", tmpdir, "build", "test.py"])
-        assert result.exit_code == 0, result.stdout
-        cache_path = tmpdir / ".wily"
-        assert cache_path.exists()
-        index_path = tmpdir / ".wily" / "filesystem" / "index.json"
-        assert index_path.exists()
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", tmpdir, "--cache", cache_path, "build", "test.py"]
+    )
+    assert result.exit_code == 0, result.stdout
+    cache_path = pathlib.Path(cache_path)
+    assert cache_path.exists()
+    index_path = cache_path / "filesystem" / "index.json"
+    assert index_path.exists()
 
 
-def test_build_invalid_path(tmpdir):
+def test_build_custom_cache(tmpdir):
+    """
+    Test that build defaults to filesystem in a non-git directory with custom cache path.
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", tmpdir, "--cache", tmpdir / ".wily", "build", "test.py"]
+    )
+    assert result.exit_code == 0, result.stdout
+    cache_path = tmpdir / ".wily"
+    assert cache_path.exists()
+    index_path = cache_path / "filesystem" / "index.json"
+    assert index_path.exists()
+    assert not pathlib.Path(generate_cache_path(tmpdir)).exists()
+
+
+def test_build_invalid_path():
     """
     Test that build fails with a garbage path
     """
-    with patch("wily.logger") as logger:
-        runner = CliRunner()
-        result = runner.invoke(main.cli, ["--path", "/fo/v/a", "build", "test.py"])
-        assert result.exit_code == 1, result.stdout
+    runner = CliRunner()
+    result = runner.invoke(main.cli, ["--path", "/fo/v/a", "build", "test.py"])
+    assert result.exit_code == 1, result.stdout
 
 
 def test_build_no_target(tmpdir):
     """
     Test that build fails with no target
     """
-    with patch("wily.logger") as logger:
-        runner = CliRunner()
-        result = runner.invoke(main.cli, ["--path", tmpdir, "build"])
-        assert result.exit_code == 2, result.stdout
+    runner = CliRunner()
+    result = runner.invoke(main.cli, ["--path", tmpdir, "build"])
+    assert result.exit_code == 2, result.stdout
 
 
 def test_build_crash(tmpdir):
     """
-    Test that build works in a basic repository.
+    Simulate a runtime error in the build.
     """
     repo = Repo.init(path=tmpdir)
     tmppath = pathlib.Path(tmpdir)
@@ -61,16 +82,15 @@ def test_build_crash(tmpdir):
     with open(tmppath / "test.py", "w") as test_txt:
         test_txt.write("import abc")
 
-    with open(tmppath / ".gitignore", "w") as test_txt:
-        test_txt.write(".wily/")
-
     index = repo.index
-    index.add(["test.py", ".gitignore"])
+    index.add(["test.py"])
 
     author = Actor("An author", "author@example.com")
     committer = Actor("A committer", "committer@example.com")
 
     index.commit("basic test", author=author, committer=committer)
+    repo.close()
+
     import wily.commands.build
 
     with patch.object(
@@ -81,20 +101,8 @@ def test_build_crash(tmpdir):
         assert bar_finish.called_once
         assert result.exit_code == 1, result.stdout
 
-    with patch("wily.commands.build.logger") as logger:
-        logger.level = "DEBUG"
-        with patch.object(
-            wily.commands.build.Bar, "finish", side_effect=RuntimeError("arggh")
-        ) as bar_finish:
-            runner = CliRunner()
-            result = runner.invoke(
-                main.cli, ["--debug", "--path", tmpdir, "build", "test.py"]
-            )
-            assert bar_finish.called_once
-            assert result.exit_code == 1, result.stdout
 
-
-def test_build(tmpdir):
+def test_build(tmpdir, cache_path):
     """
     Test that build works in a basic repository.
     """
@@ -105,33 +113,31 @@ def test_build(tmpdir):
     with open(tmppath / "test.py", "w") as test_txt:
         test_txt.write("import abc")
 
-    with open(tmppath / ".gitignore", "w") as test_txt:
-        test_txt.write(".wily/")
-
     index = repo.index
-    index.add(["test.py", ".gitignore"])
+    index.add(["test.py"])
 
     author = Actor("An author", "author@example.com")
     committer = Actor("A committer", "committer@example.com")
 
     commit = index.commit("basic test", author=author, committer=committer)
+    repo.close()
 
-    with patch("wily.logger") as logger:
-        runner = CliRunner()
-        result = runner.invoke(
-            main.cli, ["--debug", "--path", tmpdir, "build", "test.py"]
-        )
-        assert result.exit_code == 0, result.stdout
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli,
+        ["--debug", "--path", tmpdir, "--cache", cache_path, "build", "test.py"],
+    )
+    assert result.exit_code == 0, result.stdout
 
-    cache_path = tmpdir / ".wily"
+    cache_path = pathlib.Path(cache_path)
     assert cache_path.exists()
-    index_path = tmpdir / ".wily" / "git" / "index.json"
+    index_path = cache_path / "git" / "index.json"
     assert index_path.exists()
-    rev_path = tmpdir / ".wily" / "git" / commit.name_rev.split(" ")[0] + ".json"
+    rev_path = cache_path / "git" / (commit.name_rev.split(" ")[0] + ".json")
     assert rev_path.exists()
 
 
-def test_build_twice(tmpdir):
+def test_build_twice(tmpdir, cache_path):
     """
     Test that build works when run twice.
     """
@@ -141,10 +147,9 @@ def test_build_twice(tmpdir):
     # Write a test file to the repo
     with open(tmppath / "test.py", "w") as test_txt:
         test_txt.write("import abc")
-    with open(tmppath / ".gitignore", "w") as test_txt:
-        test_txt.write(".wily/")
+
     index = repo.index
-    index.add(["test.py", ".gitignore"])
+    index.add(["test.py"])
 
     author = Actor("An author", "author@example.com")
     committer = Actor("A committer", "committer@example.com")
@@ -152,14 +157,17 @@ def test_build_twice(tmpdir):
     commit = index.commit("basic test", author=author, committer=committer)
 
     runner = CliRunner()
-    result = runner.invoke(main.cli, ["--debug", "--path", tmpdir, "build", "test.py"])
+    result = runner.invoke(
+        main.cli,
+        ["--debug", "--path", tmpdir, "--cache", cache_path, "build", "test.py"],
+    )
     assert result.exit_code == 0, result.stdout
 
-    cache_path = tmpdir / ".wily"
+    cache_path = pathlib.Path(cache_path) / "git"
     assert cache_path.exists()
-    index_path = tmpdir / ".wily" / "git" / "index.json"
+    index_path = cache_path / "index.json"
     assert index_path.exists()
-    rev_path = tmpdir / ".wily" / "git" / commit.name_rev.split(" ")[0] + ".json"
+    rev_path = cache_path / (commit.name_rev.split(" ")[0] + ".json")
     assert rev_path.exists()
 
     # Write a test file to the repo
@@ -169,17 +177,17 @@ def test_build_twice(tmpdir):
     index.add(["test.py"])
 
     commit2 = index.commit("basic test", author=author, committer=committer)
+    repo.close()
 
     result = runner.invoke(main.cli, ["--debug", "--path", tmpdir, "build", "test.py"])
     assert result.exit_code == 0, result.stdout
 
-    cache_path = tmpdir / ".wily"
     assert cache_path.exists()
-    index_path = tmpdir / ".wily" / "git" / "index.json"
+    index_path = cache_path / "index.json"
     assert index_path.exists()
-    rev_path = tmpdir / ".wily" / "git" / commit.name_rev.split(" ")[0] + ".json"
+    rev_path = cache_path / (commit.name_rev.split(" ")[0] + ".json")
     assert rev_path.exists()
-    rev_path2 = tmpdir / ".wily" / "git" / commit2.name_rev.split(" ")[0] + ".json"
+    rev_path2 = cache_path / (commit2.name_rev.split(" ")[0] + ".json")
     assert rev_path2.exists()
 
 
@@ -188,6 +196,7 @@ def test_build_no_commits(tmpdir):
     Test that build fails cleanly with no commits
     """
     repo = Repo.init(path=tmpdir)
+    repo.close()
 
     runner = CliRunner()
     result = runner.invoke(
@@ -211,9 +220,11 @@ def test_build_dirty_repo(builddir):
 
 def test_build_no_git_history(tmpdir):
     repo = Repo.init(path=tmpdir)
+    repo.close()
+
     with patch("wily.logger") as logger:
         runner = CliRunner()
-        result = runner.invoke(main.cli, ["--path", tmpdir, "build", "src/test.py"])
+        result = runner.invoke(main.cli, ["--path", tmpdir, "build", _path])
         assert result.exit_code == 1, result.stdout
 
 
@@ -221,17 +232,18 @@ archivers = {name for name in ALL_ARCHIVERS.keys()}
 
 
 @pytest.mark.parametrize("archiver", archivers)
-def test_build_archiver(gitdir, archiver):
+def test_build_archiver(gitdir, archiver, cache_path):
     """
     Test the build against each type of archiver
     """
     with patch("wily.logger") as logger:
         runner = CliRunner()
         result = runner.invoke(
-            main.cli, ["--path", gitdir, "build", "src/test.py", "-a", archiver]
+            main.cli,
+            ["--path", gitdir, "--cache", cache_path, "build", _path, "-a", archiver],
         )
         assert result.exit_code == 0, result.stdout
-        cache_path = gitdir / ".wily"
+        cache_path = pathlib.Path(cache_path)
         assert cache_path.exists()
-        index_path = gitdir / ".wily" / archiver / "index.json"
+        index_path = cache_path / archiver / "index.json"
         assert index_path.exists()