+++ test/integration/test_report.py
@@ -1,11 +1,18 @@
+import sys
+
 from click.testing import CliRunner
+from pathlib import Path
 
 import wily.__main__ as main
 
+_path = "src\\test.py" if sys.platform == "win32" else "src/test.py"
 
-def test_report_no_cache(tmpdir):
+
+def test_report_no_cache(tmpdir, cache_path):
     runner = CliRunner()
-    result = runner.invoke(main.cli, ["--path", tmpdir, "report", "src/test.py"])
+    result = runner.invoke(
+        main.cli, ["--path", tmpdir, "--cache", cache_path, "report", _path]
+    )
     assert result.exit_code == 1, result.stdout
 
 
@@ -16,14 +23,7 @@ def test_report(builddir):
     runner = CliRunner()
     result = runner.invoke(
         main.cli,
-        [
-            "--path",
-            builddir,
-            "report",
-            "src/test.py",
-            "raw.multi",
-            "maintainability.rank",
-        ],
+        ["--path", builddir, "report", _path, "raw.multi", "maintainability.rank"],
     )
     assert result.exit_code == 0, result.stdout
     assert "Not found" not in result.stdout
@@ -40,7 +40,7 @@ def test_report_granular(builddir):
             "--path",
             builddir,
             "report",
-            "src/test.py:function1",
+            _path + ":function1",
             "cyclomatic.complexity",
             "-n",
             1,
@@ -56,7 +56,7 @@ def test_report_not_found(builddir):
     """
     runner = CliRunner()
     result = runner.invoke(
-        main.cli, ["--path", builddir, "report", "src/test1.py", "raw.loc"]
+        main.cli, ["--path", builddir, "report", "test1.py", "raw.loc"]
     )
     assert result.exit_code == 0, result.stdout
     assert "Not found" in result.stdout
@@ -67,7 +67,7 @@ def test_report_default_metrics(builddir):
     Test that report works with default metrics
     """
     runner = CliRunner()
-    result = runner.invoke(main.cli, ["--path", builddir, "report", "src/test.py"])
+    result = runner.invoke(main.cli, ["--path", builddir, "report", _path])
     assert result.exit_code == 0, result.stdout
     assert "Not found" not in result.stdout
 
@@ -88,8 +88,7 @@ def test_report_with_message(builddir):
     """
     runner = CliRunner()
     result = runner.invoke(
-        main.cli,
-        ["--path", builddir, "report", "src/test.py", "raw.multi", "--message"],
+        main.cli, ["--path", builddir, "report", _path, "raw.multi", "--message"]
     )
     assert result.exit_code == 0, result.stdout
     assert "basic test" in result.stdout
@@ -103,7 +102,7 @@ def test_report_high_metric(builddir):
     """
     runner = CliRunner()
     result = runner.invoke(
-        main.cli, ["--path", builddir, "report", "src/test.py", "raw.comments"]
+        main.cli, ["--path", builddir, "report", _path, "raw.comments"]
     )
     assert result.exit_code == 0, result.stdout
     assert "Not found" not in result.stdout
@@ -115,7 +114,83 @@ def test_report_low_metric(builddir):
     """
     runner = CliRunner()
     result = runner.invoke(
-        main.cli, ["--path", builddir, "report", "src/test.py", "maintainability.mi"]
+        main.cli, ["--path", builddir, "report", _path, "maintainability.mi"]
     )
     assert result.exit_code == 0, result.stdout
+    assert "Not found" not in result.stdout
+
+
+def test_report_html_format(builddir):
+    """
+    Test that report works with HTML as format
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", builddir, "report", _path, "--format", "HTML"]
+    )
+    path = Path().cwd()
+    path = path / "wily_report" / "index.html"
+
+    assert path.exists()
+    assert "<html>" in path.read_text()
+    assert result.exit_code == 0, result.stdout
+    assert "Not found" not in result.stdout
+
+
+def test_report_html_format_target_folder(builddir):
+    """
+    Test that report works with HTML as format
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", builddir, "report", _path, "--format", "HTML", "-o", "foo"]
+    )
+    path = Path().cwd()
+    path = path / "foo" / "index.html"
+
+    assert path.exists()
+    assert "<html>" in path.read_text()
+    assert result.exit_code == 0, result.stdout
+    assert "Not found" not in result.stdout
+
+
+def test_report_html_format_target_file(builddir):
+    """
+    Test that report works with HTML as format
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", builddir, "report", _path, "--format", "HTML", "-o", "foo/bar.html"]
+    )
+    path = Path().cwd()
+    path = path / "foo" / "bar.html"
+
+    assert path.exists()
+    assert "<html>" in path.read_text()
+    assert result.exit_code == 0, result.stdout
+    assert "Not found" not in result.stdout
+
+
+def test_report_console_format(builddir):
+    """
+    Test that report works with console as format
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli, ["--path", builddir, "report", _path, "--format", "CONSOLE"]
+    )
+    assert result.exit_code == 0, result.stdout
+    assert "Not found" not in result.stdout
+
+
+def test_report_not_existing_format(builddir):
+    """
+    Test that report works with non-existing format
+    """
+    runner = CliRunner()
+    result = runner.invoke(
+        main.cli,
+        ["--path", builddir, "report", _path, "--format", "non-existing"],
+    )
+    assert result.exit_code == 2, result.stdout
     assert "Not found" not in result.stdout