+++ wily/config.py
@@ -6,10 +6,11 @@ TODO : Better utilise default values and factory in @dataclass to replace DEFAUL
  and replace the logic in load() to set default values.
  MODULE:0-3
 """
-
+from functools import lru_cache
 import configparser
 import logging
 import pathlib
+import hashlib
 from dataclasses import dataclass, field
 from typing import Any, List
 
@@ -18,8 +19,24 @@ from wily.archivers import ARCHIVER_GIT
 
 logger = logging.getLogger(__name__)
 
-""" The default path name to the cache """
-DEFAULT_CACHE_PATH = ".wily"
+
+@lru_cache(maxsize=128)
+def generate_cache_path(path):
+    """
+    Generate a reusable path to cache results.
+
+    Will use the --path of the target and hash into
+    a 9-character directory within the HOME folder.
+
+    :return: The cache path
+    :rtype: ``str``
+    """
+    logger.debug(f"Generating cache for {path}")
+    sha = hashlib.sha1(str(path).encode()).hexdigest()[:9]
+    HOME = pathlib.Path.home()
+    cache_path = str(HOME / ".wily" / sha)
+    logger.debug(f"Cache path is {cache_path}")
+    return cache_path
 
 
 @dataclass
@@ -34,8 +51,6 @@ class WilyConfig(object):
     archiver: Any
     path: str
     max_revisions: int
-    skip_ignore_check: bool = False
-    cache_path: str = DEFAULT_CACHE_PATH
     targets: List[str] = None
     checkout_options: dict = field(default_factory=dict)
 
@@ -43,6 +58,20 @@ class WilyConfig(object):
         """Clone targets as a list of path."""
         if self.targets is None or "":
             self.targets = [self.path]
+        self._cache_path = None
+
+    @property
+    def cache_path(self):
+        """Path to the cache."""
+        if not self._cache_path:
+            self._cache_path = generate_cache_path(pathlib.Path(self.path).absolute())
+        return self._cache_path
+
+    @cache_path.setter
+    def cache_path(self, value):
+        """Override the cache path."""
+        logger.debug(f"Setting custom cache path to {value}")
+        self._cache_path = value
 
 
 # Default values for Wily
@@ -52,6 +81,7 @@ DEFAULT_OPERATORS = {
     operators.OPERATOR_RAW.name,
     operators.OPERATOR_MAINTAINABILITY.name,
     operators.OPERATOR_CYCLOMATIC.name,
+    operators.OPERATOR_HALSTEAD.name,
 }
 
 """ The name of the default archiver """