+++ wily/operators/maintainability.py
@@ -6,12 +6,24 @@ Measures the "maintainability" using the Halstead index.
 MODULE:3-3
 """
 import statistics
+from collections import Counter
 
 import radon.cli.harvest as harvesters
 from radon.cli import Config
 
 from wily import logger
 from wily.operators import BaseOperator, MetricType, Metric
+
+
+def mode(data):
+    """
+    Return the modal value of a iterable with discrete values.
+    
+    If there is more than 1 modal value, arbritrarily return the first top n.
+    """
+    c = Counter(data)
+    mode, freq = c.most_common(1)[0]
+    return mode
 
 
 class MaintainabilityIndexOperator(BaseOperator):
@@ -29,13 +41,7 @@ class MaintainabilityIndexOperator(BaseOperator):
     }
 
     metrics = (
-        Metric(
-            "rank",
-            "Maintainability Ranking",
-            str,
-            MetricType.Informational,
-            statistics.mode,
-        ),
+        Metric("rank", "Maintainability Ranking", str, MetricType.Informational, mode),
         Metric(
             "mi", "Maintainability Index", float, MetricType.AimLow, statistics.mean
         ),